+++ app/nexus/internal/state/backend/sqlite/key.md.go
@@ -3,15 +3,3 @@
 // \\\\\\\ SPDX-License-Identifier: Apache-2.0
 
 package sqlite
-
-type DatabaseConfigKey string
-
-const (
-	KeyDataDir                DatabaseConfigKey = "data_dir"
-	KeyDatabaseFile           DatabaseConfigKey = "database_file"
-	KeyJournalMode            DatabaseConfigKey = "journal_mode"
-	KeyBusyTimeoutMs          DatabaseConfigKey = "busy_timeout_ms"
-	KeyMaxOpenConns           DatabaseConfigKey = "max_open_conns"
-	KeyMaxIdleConns           DatabaseConfigKey = "max_idle_conns"
-	KeyConnMaxLifetimeSeconds DatabaseConfigKey = "conn_max_lifetime_seconds"
-)
